"The following changes have been implemented.
1. After speaking with the team, I realized that some properties need
to be split out from the setCoreProperties function of common.js. This
function is typically used with the outermost element associated with
the widget and properties like id, style, and class belong on this
element. There is no problem with simple widgets like image; however,
properties such as title, dir, and lang may belong to an inner element
(e.g., an input for a checkbox, text field, etc.). Therefore, I
created another function, named setCommonProperties.

In addition, I noticed another minor issue with the widgets. The
label, staticText, and image update all properties when the setProps
function is invoked. For example, the label always updates the error
image regardless if props.errorImage is null or not. (We're testing
the _props object, so properties are always being set.) Further, when
the subcomponent is updated, we're providing _props.errorImage as the
argument to setProps. Providing all default properties forces the
entire widget to be redrawn unnecessarily.

That said, we should be able to update just the src property of an
image without redrawing the entire widget. Therefore, I modified the
existing widget code a bit. Now updates happen only when new
properties are provided.

2. Since the setProps function shall only update a subset of
properties, keeping a private _props variable on the DOM node is no
longer necessary. Considering that the widget object already contains
these properties, I believe we should remove the _props variable from
the DOM node and update the widget object instead. Thus, we don't have
copies of properties in two places.

That said, I still want to place public functions on the DOM
node. However, we don't need to add our private functions there. This
was useful mainly because the _props variable was there as well. Thus,
we'll add private functions (if any) to the widget object and keep the
DOM node simple. No need to add the "_" prefix for functions of the
widget object because that's already considered private.

3. Instead calling setProps, from the fillInTemplate function, with a
newly created object like so:

    // Set properties.
    this.domNode.setProps({
        alt: this.alt,
        align: this.align,
        className: this.className,
        ...
    });

We can call this.domNode.setProps(this), instead. Note that there may
be a couple extra properties in the widget object (e.g., Dojo attach
points), but it saves us from creating one more object. All the
properties are stored in the widget, anyway."

svn path=/trunk/; revision=85


Former-commit-id: 7a2637da1a518021c10da3ff341a2d104fc409c5