"When components implement the Widget interface, it affects how widget children are rendered. That is, if a component parent also implements Widget, the JavaScript used to instantiate the client-side widget is not output. Instead, only JSON properties are output. Not only is the JavaScript unnecessary for a widget child, but it allows subcomponents properties to be updated client-side.
That said, Venky encountered an issue while implementing the Widget
interface for Hyperlink. There are a few components that extend
Hyperlink, but are not client-side widgets themselves. Specifically,
this is a problem for AlarmStatus, JobStatus, etc. For example, if an
Image widget is rendered as a child of AlarmStatus, the image will not
render correctly. This is because its AlarmStatus parent also inherits
the Widget interface. In this scenario, the JavaScript required to
instantiate the image widget is omitted.

Instead of rendering components based on the Widget interface, I've
modified RendererBase to test the renderer assigned to the component
parent. In the case of AlarmStatus, its renderer does not extend
RendererBase. Therefore, the image widget will output the JavaScript
necessary to instantiate the client-side image widget. This means that
we do not need to create client-side widgets for AlarmStatus,
JobStaus, etc., as Venky feared.
"

svn path=/trunk/; revision=90


Former-commit-id: bd7456d2bc4f953342c4a5c0ca3995f21174f78a