Issue number: 67 Submitted by: larkula Reviewed by: Self
Fixed the property editor issues for "sortField and "lookin" properties in FileChooser.

Removed the editorClassName for "sortField" in the component class, and applied the appropriate property domain in the bean info class instead.

Also, restored declaration for the "lookin" property's editorClassName.
This was inadvertently removed as part of the fix for issue # 7.

svn path=/trunk/; revision=91


Former-commit-id: 4681f13646ad9c21a8de68d0f250827ba005bf0c