"At one time, each widget renderer output their own JavaScript include(s) in the page. However, it's important that all JavaScript include statements are added to the page before instantiating the widget. Omitting some details, this meant that all child components needed to be rendered before outputting JavaScript includes in the page. Thus, the encodeBegin and endcodeEnd methods were not used in RendererBase.
Fortunately, we found a better approach and this functionality was
removed some time ago. JavaScript includes are now defined via the
abstract getModules() method. In short, a modules property is added to
the JSON object representing the includes required for each
component. This allows the modules (i.e., JavaScript includes) to be
added client-side via Dojo. Thus, JavaScript includes are added
exactly once when the widget is instantiated.

That said, I've modified RendererBase to use the encodeBegin and
encodeChildren methods, again. Now encodeBegin and encodeEnd simply
output the JavaScript used to instantiate a widget. There is no
change to the widget renderers themselves.
"

svn path=/trunk/; revision=96


Former-commit-id: 6e4a84e741a8ca0248429faef7f260daf9a0e48d