"After the getProps() function was reviewed for button, I went through the existing widgets (e.g., image, label, etc.) and updated those as well.  After this put back, all the widgets will be implemented similarly and follow the latest architecture.
I also added one small change to HeadRenderer.java. In order to debug
JavaScript, we're forced to edit the JSP page and set the debug
attribute of the head tag. Editing the JSP can be a pain, especially
if you're debugging a web app not on your local machine. Therefore, if
the "debug" query paramter is added to a URL, debugging is turned on
without editing the JSP page. For example:

http://localhost:8080/testapp/faces/index.jsp?debug
"

svn path=/trunk/; revision=112


Former-commit-id: f66d04496755d637787704d650483ae0eb26ae4f