IN=AUTO update dynafaces jars and some virtual forms designtime code
This diff includes updating the dynamic faces jars to the latest available.

In VWP, we are adding some extensions to the designtime api in a new module called designtimeext. These extensions include interfaces that abstract the painting of component groups in the designer, such as is done for virtual forms. So this integration modifies FormDesignInfo and some of the virtual form designtime customizers to use the designtimeext interfaces. We are mainly doing this to support similar visual functionality in the designer for Ajax.

svn path=/trunk/; revision=159


Former-commit-id: 38eb3f7a34b9ad2d7e2c2bd8993c9026b71aa11d