"There is a problem with the drop down layout on IE. This is especially noticeable in a table action bar because the filter and action menus line wrap. With Mike's help, I was able to race this down to a <br> tag in the HTML template.
I could not remove the <br> tag from the DOM node on IE, so I added a
span instead. The span is hidden until needed and the labelOnTop is
not updatable via the setProps() function. We can always choose
another solution, but I cannot spend anymore time on this particular
issue.
"

svn path=/trunk/; revision=194


Former-commit-id: 703296bc131a6d34c1aa70834b9235b982ad817c