"I've reverted the HTML button to an HTML input element to support IE 6. I've also added instanceof checks to the widget renderers to catch casting errors.
Currently, the button widget uses an HTML button instead of the HTML
input used in 4.0. This allows any HTML element (e.g., image, text,
etc.) to be added to the body of the button tag. Not only can you
create some fancy buttons, but the widget doesn't need extra
properties (e.g., src, alt, etc.) just to create icon/submit/reset
buttons. My testing shows that this all works great on Mozilla,
FireFox, and IE 7; however, IE 6 is not working as expected. A single
button in the page submits fine, but problems occur with multiple
buttons.

The problem appears to be that IE 6 submits all name value pairs in
the page. Since each button must have a name attribute, the server
cannot distinguish which button generated the request; thus, actions
are not performed correctly. The fact that IE 6 submits all name value
pairs means that all the buttons in the page are decoded; thus,
multiple actions are queued.
"

svn path=/trunk/; revision=195


Former-commit-id: 767ffd2a392889cb402c899a70de5337ea76dc7d