"Solved an issue with the checkbox on IE 6. In the table example, I select a row and click the sort link. When the page refreshes, the checkbox is unselected even though the widget has set checked to true. I've experimented with using setAttribute Vs setting the DOM node value directly, but that wasn't the issue. What's puzzling is that the table's de/select all buttons can set the checkbox selected via the same setProps() function.
IMO, this appears to be a timing issue regarding when nodes are added
to the DOM and when IE recognizes our initialization. Probably the
same IE thread problem with using innerHTML -- we used a prototype
solution here. I was actually able to solve this particular issue by
delaying the call to setProps(). In the fillInTemplate function, I use
the following timeout. Although, the timeout is set at zero
milliseconds, it's enough to resolve the problem.

   setTimeout(webui.@THEME@.widget.checkbox.createInitCallback(this.id), 0);

That said, I've created a generic function, named initProps() that all
widgets will invoke instead of calling setProps() directly. I've add a
check for IE 6 and only selectively use the timeout; otherwise, we'll
just call setProps() directly.
"

svn path=/trunk/; revision=204


Former-commit-id: 9d8b371eaa69bf9ae91fa8ea39d7059d3f55e4a4