"It turns out that IE does not support the name attribute being added dynamically as documented at:
http://msdn.microsoft.com/workshop/author/dhtml/reference/properties/name_2.asp

In order to create an HTML element with a name attribute, the name and
value must be included when using the innerHTML property or when using
the document.createElement() function. As a work around, I included a
dummy name="forIE" attribute in the HTML template. The name value is
reset via the setProps() function.
"

svn path=/trunk/; revision=213


Former-commit-id: 27dd863656cdfad6732305629429855d69c43729