"It seems that the IE radio button fix still needs some tweaking. I was under the impression that as long as the name attribute is present when the HTML input element is created, it can be reset later. (The DOM inspector shows that the property is updated, but the radio button does not reflect this change.) It turns out that we only get one shot at setting the name attribute. After the HTML input element has been created, the attribute can never be set, again.
The solution I implemented is to use name="${this.name}" in the HTML
template. Since we're already providing the name property to the
widget, this will work fine. However, setting this property after Dojo
instantiates the widget has no affect. That is, we cannot set
this.name in the fillInTemplate function.

We try to set a default value when the name property is null, but that
has no affect on IE. The RbCBRendererBase class does output a default
name property, so we're okay there.
"

svn path=/trunk/; revision=223


Former-commit-id: 10f237c20ac8d269a79979de4bc8aba8e8c94b29