Issue number: Obtained from: Submitted by: Reviewed by: When obtaining an image's properties like "<image-key>_ALT" and "<image-key>_TITLE", the value may be a theme key which needs to be resolved to the literal text. The last changes made to "getImage" when "ResourceBundleTheme.java" was committed did not resolve the value by calling "getMessage".
This change addresses that issue and resolves the value
in case it is a theme key. If "getMesage" returns null it means the
value of "<image-key>_ALT" or "<image-key>_TITLE" may actually
be literal text and not a key, so in that case set the members
of ThemeImage "alt" and "title" to the literal values of the
"<image-key>_ALT" or "<image-key>_TITLE" properties.

svn path=/trunk/; revision=229


Former-commit-id: c8edb42b301740271473d8ae1968d0c7cfad02e7