"The radio button has really caused me some grief. On IE 6/7, the checked state of a radio button cannot be set dynamically until the input element has been added to the DOM. The only way I could set checked state was to use a timeout. Considering that checkbox has the same issue on IE 6,  I used a similar timeout for all widgets rendered on IE.
Unfortunately, there are a couple cases where this approach did not
work. For example, images do not behave well on IE unless the src
attribute is set immediately. And some examples (i.e., upload chooser)
use in-line JavaScript to manipulate buttons. If a widget has not
called setProps(), due to a timeout, the in-line JavaScirpt fails
because the widget has not been initialized.

That said, I've backed out my previous fix and implemented a different
solution. The timeout is still needed for checkbox and radiobutton,
but now it's used only to set the checked property. This allows the
rest of the widgets to be initialized while the checked state is set
eventually -- timeout is 0 milliseconds. This isn't full proof, but
it's the best I can do for now without breaking everything else.

Note that the masthead dot and separator images also display
now. Although we didn't see broken images with build #3, they were not
actually being displayed at all. You might see one separator, but
others were missing. This was due to the masthead outputting multiple
images, but using the same id for all. Every component in the page
needs a unique id; otherwise, it may be overridden.

I also fixed the decode methods for radiobutton and checkbox
groups. Although SelectorGroupRenderer no longer works with the HTML
renderers (because the input element id changed), this will allow
group components to function with the new widgets. When the new rb/cb
group widget is available, we can back out this temporary fix.
"

svn path=/trunk/; revision=230


Former-commit-id: 062e3446ee496f9944f7f81d778f27addd7c6197