"Backing out previsous change to SelectorGroupRenderer...
The check for ComplexComponent should not be in SelectorGroupRenderer.
The assumption is that the Group's  clientId is fixed at the time
the component is rendered and will be the same when the Group
component is decoded and is unique in the page.

The Group's id is placed on the rb or cb's name attribute which must
not be changed in any way in the client side radiobutton or checkbox
renderer. This value is not influenced by any labelling. It is just a
constant.
"

svn path=/trunk/; revision=234


Former-commit-id: 039c486f3b74b22aced8255685744608d8fefb3c