"Corrections for the anchor and hyperlink widgets.
1. Anchor widget ID is being modified instead of the DOM node only.
2. When user's onClick function returns false, the hyperlink submits anyway.
3. Cannot add/remove hyperlink contents. Although updating was a nice
attempt, we should keep the behavior similar to the label, where
contents are simply replaced.

Note that the "toggle image" example will need to be modified when
ImageHyperlink is available. (Considering that the full client-id
needs to be known in order to update the image, you might as well
modify the image directly.) ImageHyperlink will take care of
displaying en/disabled images, so updating contents via the hyperlink
will not be necessary.

4. Applied various Java coding conventions to indentation, docs, etc.

Note that none of this addresses the refresh functionality.
"

svn path=/trunk/; revision=247


Former-commit-id: 89060117c1bb503d3e55549e1055353d18694edf