"This allows the label to be decoupled from the text field while still updating itself during auto-validation. Although a web app developer can easily subscribe to the event topic themselves, providing a comma separated list of client ids to the text field's notify tag attribute avoids writing any JavaScript at all.
The implementation follows the approach we already use with the
refresh functionality. If you're not familiar, we created a JavaScript
closure which contains event topics, functions to publish begin/end
events, and a function to process the event itself. Although we don't
need to publish event topics for label and alert, we can still use
this approach in any widget that handles validation events. For
example, in order to invoke the refresh functionality of a widget, we
call its refresh() function. Using the same approach for validation,
the text field calls a validate() function for each widget provided via
the "notify" tag attribute.

The nice thing about this approach is that we don't need to hard code
event topics at all. Whether the event was generated by a text field
or text area, the label and alert never need to be modified -- not
even when new widgets come on line. New widgets simply need to
implement the "notify" tag attribute and call the validate() function as
needed. Best of all, each and every label widget in the page does not
need to process the same arbitrary event topics.
"

svn path=/trunk/; revision=295


Former-commit-id: f3cd06c5f4ebf46b9af78ec1e6785362c9cd0461