Fix for NetBeans VisualWebPack issu 99560.
Since the FilterCriteria class is annotated as being a component, it is included
in the faces-config file and marked as being a component. But the class does not
extend UIComponent, hence the exception.

The best solution to this problem is to refactor the annotations library into
two, one just for beans, so that design-time for classes like this
FilterCriteria can be generated without its being declared a run-time component.

Work-around for now: remove the annotations from all FilterCriteria classes, and
hand-code their BeanInfo.

svn path=/trunk/; revision=365


Former-commit-id: df966b7efab47991c97f580301173cdedac8c5a8