"I've removed the ajaxify property from the components. Instead, an ajaxify tag attribute has been added to the head, themeLinks, and portalTheme tags. If ajaxify is true (default), we will include the "webui.suntheme.widget.jsfx" module in the page as normal. When this name space is available client-side, it's a flag indicating that we can/should include the default Ajax implementation. Therefore, instead of including two files for every widget, we'll delay the include until needed (i.e., the first usage of a feature).
This will cut the number of JavaScript files downloaded in half. And
web app developers can still override the default Ajax implementation,
just not for individual components. This ensures that the Ajax
implementation used by subcomponents will not interfere with a custom
implementation.

Note that all the examples and design-time code remains the same
(i.e., no need to include an ajaxify property). A new ajaxify()
function was added to widgetBase instead. Each widget shall call this
before publishing an Ajax event, ensuring just-in-time availability."

svn path=/trunk/; revision=380


Former-commit-id: d108ce82cbf3551e4d5450bb0fabe335dde088b6