"In order to increase performance even further, we can immediately decrease the number of requests made for resources. Unless a custom HTML template is used, we can eliminate these requests by embedding a default template string in our JavaScript. Currently, each widget type downloads an HTML template file.
Basically, I've done some pre-work to allow embed template strings,
while still supporting custom HTML templates. The templateString
property takes precedence, so if a custom template is used, it must be
cleared. I'm able to clear this property by refactoring the code to
include a getTemplatePath() method in each renderer. When the
RendererBase class calls this method, I check if the theme template is
null. If the theme template path is NOT null, we set Dojo's
templateString property to an empty string -- this allows a custom
template to be used.

For my next put back, I plan to comment out the properties in the
templates.js file. Although, a couple may still be used (e.g.,
resetButton.html, imageButton). (Most names are a one-to-one mapping
between the template and widget.) If web app developers choose to
enable this via the theme, templates will be used as expected. The
other thing I need to do is to create another Java tool so I can
iterate over JavaScript files and embed the HTML templates during
build-time. In order to do this, I'll need to remove the comments
wrapping the entire JS file -- it's not used, anyway."

svn path=/trunk/; revision=419


Former-commit-id: ec85981e7f7faccd1c0a1053300561560da1954f