"This is a very simple modification for the upcoming client-side theme changes. We already introduced a getTemplatePath() function in theme.js; however, I simply made the widgets start using it.
FYI... The widgets must use either a template path or
string. Therefore, either getTemplatePath() or getTemplateString()
functions of theme.js will return null. When the widget's templatePath
property is NOT null, its templateString property is cleared via the
buildRendering function of widgetBase. For example, if the web app
developer sets the tag's htmlTemplate attribute, we want to use that
path instead of the default templateString. Clearing templateString
allows the user's template path to take precedence."

svn path=/trunk/; revision=477


Former-commit-id: da44d848276c10629b3993c518c7a4aadd43430d