"The original setProps() function is not very intuitive because there are two special cases it must handle:
1. The widget object must be extended so properties can be referenced later.
2. Some widget object properties must be cleared (e.g., we do not want
to extend label contents).

Neither one of these cases should be applied during initialization.
Therefore, we really need two functions; setProps() and
setWidgetProps(). The setProps() function has been added to widgetBase
and will automatically extend the widget object whenever properties
are set. This function will, in turn, call setWidgetProps(), which
does the bulk of the work. The setWidgetProps() function does NOT
extend the widget object and shall be called during
initialization. The setProps() function shall be overridden as needed
(e.g., in order to clear the contents property of label)."

svn path=/trunk/; revision=508


Former-commit-id: 7a4520ae09e68b6e670322cc80c0917b09bffc61