"Client side theme modifications.
- Use the client theme to define the error and required
 images if not defined in the props passed to the constructor.
 This is accomplished by implementing an "initialize" method
 in label.js.

- Add "theme" property to widgetBase.js that is mapped to
 "webui.@THEME@.theme.common" for use by widgets as
"this.theme.getMessage" etc. This produces shorter references at the
call site.

- Add "widget" property to widgetBase.js that is mapped to
 "webui.@THEME@.widget.common" for use by widgets as
"this.theme.getMessage" etc. This produces shorter references at the
call site.

This simply means that functions in the widget common module will be
accessed using the "this.widget" prefix. Of course, you can still use
webui.@THEME@.widget.common.addFragment(), but the shorthand version
is cleaner.

- Create "...widget.common.getImage" to assist widgets when creating
 default images. It takes care of providing the "widgetName"
 and "module" properties for an image. The method accepts
 an Object which is will "extend" the properties returned by
 the theme's getImage method returns the full set of properties.

- widget/LabelRenderer.java no longer references the
 "getErrorIcon" or "getRequiredIcon" methods of Label. Instead
 it just asks for the facets directly. This means that the props
 for the image widget will be rendered without image props unless
 there are error and/or required facets.

- Create "///widget.common.updateFragment() function to help reduce some
code. Instead of obtaining the widget and calling setProps() yourself,
the following code:

 // Update widget/add fragment.
 var calendarWidget = dojo.widget.byId(this.calendar.id);
 if (calendarWidget) {
     calendarWidget.setProps(props.calendar);
 } else {
     this.addFragment(this.calendarContainer, props.calendar);
 }

all reduces to:

 // Update/add fragment.
 this.widget.updateFragment(this.calendarContainer, props.calendar);

If the widget associated with props.id already exists, the widget's
setProps() function is invoked with the given props param. If the
widget does not exist, the widget object is instantiated via the
addFragment() function. The updateFragment() function also performs
some extra checking to ensure setProps() is available."

svn path=/trunk/; revision=559


Former-commit-id: eb4a07d879e3ce457e9493503e1f9219cc05fa97