Issue number:  899 (https://woodstock.dev.java.net/issues/show_bug.cgi?id=899) Obtained from: Submitted by:  Bhatt246 Reviewed by:   Rick R
Backporting the fix that was applied ot the trunk recently. Note that with
this change the theme is not truly backward compatible in version 4.1.1 of
Woodstock. If a developer can updated the theme by adding her own resources
then she'd have to update it with the changes to image.properties that are
being made checked in. Even if the developer does not perform this step the
tree will continue to work as will everything else...although with a degraded
performance.

svn path=/branches/Release41/; revision=936


Former-commit-id: 00eeaac68326e83d31989dbd06bc2213e0cd73d2