Issue number: Obtained from: Submitted by: Reviewed by:
- Issue #1119 regression:  text area of FC component displays multiple  &nbsp;

I'm not exactly sure why this is problem for IE7 and safari but changing the
&nbsp; to "\u00A0" resolves the problem. This is the unicode value for nbsp.

However there are longstanding issues with IE7 where the headers for the
file, size and date time extend past the listbox. This issue can be seen
in the 4.1.1 version of the filechooser.

I can address that in another Issue which I will file, if it hasn't been
filed. This seems to be related to the "%" width values in the CSS
for the "hr" elements or the table is not sizing correctly around the
cell containing the listbox.

svn path=/trunk/; revision=1107


Former-commit-id: f66d19d24eaed893cfa7b01ae69ad18c8143889b