This modification restores the existing start() function as a public API. The start function is a post initialization feature allowing widgets to retrieve initial data, begin polling, etc. Every widget inherits this function and there is an autoStart property to prevent createWidget from invoking this functionality. Currently, this feature is used by calendarField, login, progressbar, and table2RowGroup.
In the case of a progress bar, for example, the autoStart property
defaults to true. When the createWidget function instantiates the
widget, it will call a private _start function. This private function
tests the autoStart property and goes no further when set to
false. This allows developers to control when the widget is started.

The public start function tests a property, named _started. If the
widget has already been started, then there is no need to invoke
widget code again.

svn path=/trunk/; revision=1273


Former-commit-id: 7bb8150a121f6d4b4dc480ca5b1cb969ef6f5f96