This modification introduces a "jslib" module to create a stand alone JavaScript library. This is similar to the step we had been performing manually, but I've stripped away the files we don't use (e.g., property files, theme service, etc.). Thus, the jMaki build will use the same Woodstock library provided to the Netbeans library manager. Note that the "jmaki" module now depends on the "jslib" module.
Here are a few more changes related to jMaki.

1. Added code to obtain the browser's locale if one has not been
provided. This means we don't have to hard code the "en" locale in the
jMaki wrappers.

2. Added code to pull certain parameters from the request, if not
provided. This functionality is similar to the JSF version, which
allows us to set flags like debug mode, webuiAll, etc. for testing
purposes.

3. Modified the parseOnLoad feature to be false by default. This
eliminates the need to embed our config property in jMaki.

4. Modified the custom Dojo build and Woodstock module paths to allow
developers to use relative URLs. For example, when specifying the
bootstrap.js file in the page head.

svn path=/trunk/; revision=1332


Former-commit-id: 393559862adf99750d0525590daecba965ac1e19