Issue number: Obtained from: Submitted by: Reviewed by:
Issue 1227 - Dropdowns are misaligned in property sheets


The main issue here doesn't have anything to do with property sheet. I added
some offset to the listbox and dropdown selectors to add some space as per
a previously reported bug.

I moved the offset to the required image selector which provides the
added space when it is rendered.

Curiously I didn't see the same behavior for "listbox" which should have been
the case. As it happens I negelected to implement "_getLabelOnTopClassName"
in "listbox.js", as part of the interface to "labeledBase.js".

svn path=/trunk/; revision=1374


Former-commit-id: f4c8c7f2f9e47e817402860641247c1465a9e1d3